Add script and vscode task to automate running the app
The purpose of this commit is to make it easier for developers of B-Bay
to run the backend and the frontend. In order to to use the script and
vscode task, click Ctrl + Shift + P, type 'Tasks: Configure Default Build
Task', click enter, and then enter on the task titled "Start App". Once
this is done, you should be able to run the bbay frontend/backend using
the command Ctrl + Shift + B (it might ask you to configure git bash as
your defualt trmeinal to use."